9/10/2024 (1)
-pagination problem solved for student applicant view page
-pagination problme found during testing where"
StudyPal Test Case
StudyPal Test Case
100%
10
D10

I dunno how to explain this so I am going to try to explain it as clearly as possible. Basically, if I were to delete a file, and then I clicked on the arrow button below, then i upload the file again, the file doesnt show up in the table, only when i clicked on the back arrow button will it appear suddenly, at the same time, the arrow button will become dysfunctional.

I dunno how to explain this so I am going to try to explain it as clearly as possible. Basically, if I were to delete a file, and then I clicked on the arrow button below, then i upload the file again, the file doesnt show up in the table, only when i clicked on the back arrow button will it appear suddenly, at the same time, the arrow button will become dysfunctional."
solved for othercert doc and academic transcript